requirements.txt
streamlit==1.35.0
pandas==2.2.2
openai==1.30.1
python-dateutil==2.9.0
google-auth
google-api-python-client
